Move getAllUsagePlans to a "shared" file
